drm/panthor: Add the devfreq logical block
Every thing related to devfreq in placed in panthor_devfreq.c, and
helpers that can be called by other logical blocks are exposed through
panthor_devfreq.h.

This implementation is loosely based on the panfrost implementation,
the only difference being that we don't count device users, because
the idle/active state will be managed by the scheduler logic.

v2:
- Added in v2

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>