arm64: dts: rockchip: Add Turing RK1 SoM support
The Turing RK1 is an upcoming RK3588-based SoM from Turing Machines,
designed on the Jetson SO-DIMM form factor and meant to be compatible
with most Jetson carrier boards (but especially the Turing Pi 2 cluster
board from the same vendor). It has the typical I/O you'd expect from
a Jetson board, including:

- USB-OTG
- Two UARTs (one for console, one auxiliary)
- PCI Express (2.0 x1 + 3.0 x4)
- Gigabit Ethernet
- On-board eMMC
- PWM fan w/ tach
- HDMI and MIPI DSI [1]
- Miscellaneous GPIO, I²C, SPI lines [1]

Beyond that, it is pretty similar to the RK3588 EVB (in terms of PMICs,
etc).

This patch introduces a devicetree to enable most[1] of the SoM I/O,
and defines a cooling map to control the fan.

[1] These peripherals are not addressed with this patch.

Signed-off-by: Sam Edwards <CFSworks@gmail.com>